remove useless & rename file
