+++ bin/json2update
@@ -8,22 +8,6 @@ var program = require('commander');
 var _ = require('lodash');
 var jsonStream = require("json-stream");
 
-// Get command line arguments, now is form standard input & output
-program.version('0.1.0')
-  .option('-H, --host <Elastic Host>', 'Specify Elasticsearch host and port, default: localhost:9200', 'localhost:9200')
-  .option('-D, --debug', 'Enable Debugging')
-  .option('-i, --indices <Elastic indices>', 'Specify indices')
-  .option('-t, --type <Elastic type>', 'Specify type')
-  .option('--api <version>', 'Specify Elasticsearch API version, such as, 1.1, 1.2, 1.3, default: 1.4.', '1.4')
-  .parse(process.argv);
-
-var esLogLevel = 'warning';
-if (program.debug) {
-  var debug = require('debug');
-  debug.enable('elastic-write-pipe:debug');
-  esLogLevel = 'debug';
-}
-
 //use stream, jsonAddField to pipe data to standard output
 var lambdaUtil = require('../lambda-util');
 var stream = jsonStream();
